refactor: move DepartmentStatus and DepartmentType enums to a separate file
